Shopping_Application : ProductListFragment commit

+++ Shopping_Application/app/src/main/java/ows/kotlinstudy/shopping_application/data/response/ProductResponse.kt
@@ -17,7 +17,7 @@ data class ProductResponse(
             id = id.toLong(),
             createdAt = Date(createdAt),
             productName = productName,
-            productPrice = productPrice.toInt(),
+            productPrice = productPrice.toDouble().toInt(),
             productImage = productImage,
             productType = productType,
             productIntroductionImage = productIntroductionImage
